+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Fixed a bug that would stop preview frames from being retried if they failed even once.
 - Hotfix to remove double division by exposure time when creating master darks
 - Fixed bug that prevented calibration comparison being run on skyflats
+- Fixed image class null case 
 
 0.9.3 (2018-05-10)
 ------------------